requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.10
dj-database-url==0.5.0
Django==4.0.1
django-on-heroku==1.1.2
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
lxml==4.7.1
multitasking==0.0.10
numpy==1.22.1
pandas==1.3.5
psycopg2-binary==2.9.3
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
whitenoise==5.3.0
yfinance==0.1.69
